Rework nightly workflow into more general release workflow (#48)
This PR adds a 3rd trigger to the nightly release workflow in addition
to the current manual and schedule triggers. Now, when a release is
published, build will be created and uploaded to the release, using the
tag as a version name. Currently, macOS builds are excluded from this,
since the ones created by this workflow aren't packaged properly for
actual releases yet.

There's also one final-really-for-real-this-time fix for the update info
in the appimages.

Here's an example of the assets uploaded to a release tagged `0.9.1`
https://github.com/electricbrass/Cacoco/releases/tag/0.9.1